+++ x/epoching/types/msg.go
@@ -8,6 +8,8 @@ import (
 // ensure that these message types implement the sdk.Msg interface
 var (
 	_ sdk.Msg = &MsgWrappedDelegate{}
+	_ sdk.Msg = &MsgWrappedEditValidator{}
+	_ sdk.Msg = &MsgWrappedStakingUpdateParams{}
 	_ sdk.Msg = &MsgWrappedUndelegate{}
 	_ sdk.Msg = &MsgWrappedBeginRedelegate{}
 	_ sdk.Msg = &MsgWrappedCancelUnbondingDelegation{}
@@ -17,6 +19,13 @@ var (
 // NewMsgWrappedDelegate creates a new MsgWrappedDelegate instance.
 func NewMsgWrappedDelegate(msg *stakingtypes.MsgDelegate) *MsgWrappedDelegate {
 	return &MsgWrappedDelegate{
+		Msg: msg,
+	}
+}
+
+// NewMsgWrappedEditValidator creates a new MsgWrappedDelegate instance.
+func NewMsgWrappedEditValidator(msg *stakingtypes.MsgEditValidator) *MsgWrappedEditValidator {
+	return &MsgWrappedEditValidator{
 		Msg: msg,
 	}
 }
@@ -38,6 +47,13 @@ func NewMsgWrappedBeginRedelegate(msg *stakingtypes.MsgBeginRedelegate) *MsgWrap
 // NewMsgWrappedCancelUnbondingDelegation creates a new MsgWrappedCancelUnbondingDelegation instance.
 func NewMsgWrappedCancelUnbondingDelegation(msg *stakingtypes.MsgCancelUnbondingDelegation) *MsgWrappedCancelUnbondingDelegation {
 	return &MsgWrappedCancelUnbondingDelegation{
+		Msg: msg,
+	}
+}
+
+// NewMsgWrappedStakingUpdateParams creates a new MsgWrappedStakingUpdateParams instance.
+func NewMsgWrappedStakingUpdateParams(msg *stakingtypes.MsgUpdateParams) *MsgWrappedStakingUpdateParams {
+	return &MsgWrappedStakingUpdateParams{
 		Msg: msg,
 	}
 }